fix(dashboard): Fix merge conflicts
