+++ apps/dashboard/src/components/workflow-editor/steps/email/variables/variables.ts
@@ -1,4 +1,5 @@
 import { Variable, Variables } from '@maily-to/core/extensions';
+import { DYNAMIC_PATH_ROOTS, DYNAMIC_STEP_NAME_ROOT_REGEX } from '@/utils/constants';
 import type { Editor as TiptapEditor } from '@tiptap/core';
 
 export enum VariableFrom {
@@ -95,10 +96,35 @@ const getRepeatBlockEachVariable = (editor: TiptapEditor): Array<Variable> => {
 };
 
 const dedupAndSortVariables = (variables: Array<Variable>, query: string): Array<Variable> => {
-  const filteredVariables = variables.filter((variable) => variable.name.toLowerCase().includes(query.toLowerCase()));
+  // Filter variables that match the query
+  let filteredVariables = variables.filter((variable) => variable.name.toLowerCase().includes(query.toLowerCase()));
 
+  if (filteredVariables.length === 0 && variables.length > 0) {
+    /**
+     * When typing a variable that is not present in the list,
+     * we need to check if the variable is a dynamic path.
+     * If it is, we need to add the dynamic path values to the list.
+     * For example, if the user types "hello" and there are dynamic paths like "steps.x.events[n].payload",
+     * we need to add "steps.x.events[n].payload.hello", "payload.hello" , "subscriber.data.hello" to the list.
+     * This is done to allow the user to create dynamic paths.
+     */
+    const dynamicStepNames = variables
+      .map((entry) => entry.name.match(DYNAMIC_STEP_NAME_ROOT_REGEX))
+      .filter((match): match is RegExpMatchArray => match !== null)
+      .map((match) => `${match[0]}.`);
+
+    filteredVariables = [...DYNAMIC_PATH_ROOTS, ...dynamicStepNames].map((value) => {
+      return {
+        name: value + query.trim(),
+        required: false,
+      };
+    });
+  }
+
+  // Deduplicate based on name property
   const uniqueVariables = Array.from(new Map(filteredVariables.map((item) => [item.name, item])).values());
 
+  // Sort variables: exact matches first, then starts with query, then alphabetically
   return uniqueVariables.sort((a, b) => {
     const aExactMatch = a.name.toLowerCase() === query.toLowerCase();
     const bExactMatch = b.name.toLowerCase() === query.toLowerCase();
